Bloodlines implemented into the history files
All bloodlines that have been scripted are now integrated into the history. The only additional change is that Malekith is now listed as the son of Aenarion for the purpose of bloodline acquisition, however it seems this does some funny things to succession. If necessary for balance that can be easily changed back, or the bloodline can just as easily be set as inactive for those who are not either Asur or Cadai.