Update DR_chaos_dwarfs_events.txt
Currently if the decay event fires for the Dawi Zhar slave buildings no matter what level they are at, they get cut down to the tier 1 building. looking at the code I interpreted that it was not intended to do that, and instead it was intended to only go down 1 level upon decaying. I've added additional checks since CK2 buildings are cumulative and I'm fairly certain that having tier 4 of a building means also having tier 2 and 3 of said building, which then hits the current triggers for destroying tier 4, 3, and 2 of the building. With these checks it shouldn't destroy everything down to tier 1 of the buildings if you have a higher tier building.