Fixed the Nerge event for Beastmen
Added back the event file from Vanilla and so the event works when taken. Modified the event decision potential so it can only be taken by independent rulers since the event is normally only able to be taken by independent rulers. Will update discord on testing to make sure it works smoothly.